refactor: merge update-agent-channel into update-agent
Consolidate duplicated helpers (gatewayDir, configPath, loadConfig,
saveConfig, createRl, prompt) and move all channel management logic
(detectConnectedChannels, removeChannel, runMenu, setupChannel) into
update-agent.ts. Delete scripts/update-agent-channel.ts and remove
the make target. Update test import accordingly.
